movie-recommender/requirements.txt
pandas
scikit-surprise
scikit-learn
streamlit
matplotlib
